Split page move to component
