+++ framerail/src/routes/[slug]/[...extra]/index.ts
@@ -1,0 +1,1 @@
+export { default as MovePane } from "./MovePane.svelte"
